"Implementation of Lombok in the created classes"
