+++ src/main/java/co/edu/udes/backend/models/Message.java
@@ -1,4 +1,12 @@
 package co.edu.udes.backend.models;
 
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@RequiredArgsConstructor
 public class Message {
 }
